refactored offer page, solved a map warning
